PK score bar issue fixes, custom error handler added

+++ Sources/IsometrikStream/Config/APIClient/Endpoints/PK/Models/PKBattleWinnerModel.swift
@@ -25,8 +25,8 @@ public struct ISM_PK_WinnerModel: Codable {
     public private(set) var winnerId: String?
     public private(set) var percentageOfCoinsFirstUser: Double?
     public private(set) var percentageOfCoinsSecondUser: Double?
-    public private(set) var totalCoinsFirstUser: Int?
-    public private(set) var totalCoinsSecondUser: Int?
+    public private(set) var totalCoinsFirstUser: Double?
+    public private(set) var totalCoinsSecondUser: Double?
     public private(set) var status: String?
     public private(set) var winnerStreamUserId: String?
